refactor: restructure isogram functionality and tests
