+++ JS_Learning/src/1_check-isogram.js
@@ -1,0 +1,5 @@
+function isIsogram(str) {
+  str = str.toLowerCase();
+  return new Set(str).size === str.length;
+}
+module.exports = isIsogram;
